update bert and text.py
